requirements.txt
streamlit
PyYAML
streamlit-authenticator>=0.2.3
bcrypt